+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Bug fixes:
 -
 -
 -
--
+- Added support for identity impersonation when using workload identity federation (snowflakedb/gosnowflake#1652).
 -
 
 Internal changes: